# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Plus MS Auth",
    "version": "1.0.0",
    "description": "Microserviço de autenticação JWT com refresh tokens e controle de acesso por role (RBAC). Roles disponíveis: admin, moderator, user."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Desenvolvimento local"
    },
    {
      "url": "http://localhost:4566/auth",
      "description": "Ministack (LocalStack)"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registrar novo usuário",
        "description": "Cria um novo usuário com email e senha. Retorna os dados do usuário criado junto com accessToken (15min) e refreshToken (7 dias).",
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou email já registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autenticar usuário",
        "description": "Autentica com email e senha. Retorna accessToken (15min) e refreshToken (7 dias).",
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email e senha são obrigatórios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas ou usuário inativo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Dados do usuário autenticado",
        "description": "Retorna os dados do usuário extraídos do JWT. Requer Bearer token válido.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Renovar access token",
        "description": "Recebe um refreshToken válido e retorna um novo accessToken (15min).",
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token renovado com sucesso"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "eyJ..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Refresh token não informado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Encerrar sessão",
        "description": "Revoga o accessToken atual e o refreshToken do usuário. Requer Bearer token válido.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logout realizado com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@loja.com"
          },
          "password": {
            "type": "string",
            "example": "senha123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@loja.com"
          },
          "password": {
            "type": "string",
            "example": "senha123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshBody": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "uuid-longo-aqui"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid"
          },
          "email": {
            "type": "string",
            "example": "joao@loja.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-05-12T10:00:00Z"
          }
        }
      },
      "AuthTokensResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {}
          },
          "accessToken": {
            "type": "string",
            "example": "eyJ..."
          },
          "refreshToken": {
            "type": "string",
            "example": "uuid..."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "mensagem de erro"
          }
        }
      },
...
    }
  }
}